Swapped Asc/Desc on sort by Age
The sort functionality for sort by age was inversed because the database query sorts by 'birthDate' not age, meaning that an ascending sort would list the lowest year first and a descending sort would list the highest year first. The sort by age ascending link now passes a "Desc" order to the SQL query and the sort by age descending link now passes an "Asc" order to the SQL query, providing the correct logic for the user to be able to sort by age.

This is a quick fix, but a stronger solution would be to add an If statement to index.php in the $action == 'change_sort' section under $page == "all_artists" to accommodate for this reverse in sort logic and possibly add a new function in artist_db.php if necessary.